requirements/base.txt
crispy-bootstrap5==0.7
Django==4.0.3
django[argon2]==4.0.3
django-bootstrap-datepicker-plus==4.0.0
# django-crispy-forms>=1.14.0
# django-constance>=2.9.0
django-environ>=0.8.1
django-import-export>=2.8.0
django-listview-filters>=0.0.1b1.dev0
django-money>=2.1.1
django-phonenumber-field>=6.1.0
django-select2>=7.10.0
django-widget-tweaks>=1.4.12
fdfgen>=0.16.1
furl>=2.1.3
packaging>=21.3
phonenumberslite>=8.12.45
Pillow>=9.1.1
python-benedict>=0.25.2
reportlab>=3.6.10
# django-bootstrap-modal-forms>=2.2.0
shortuuid==1.0.9
# smartsheet-python-sdk>=2.105.1
